pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scdynsys"
version = "0.0.1"
authors = [
  { name="Christiaan H. van Dorp", email="chvandorp@gmail.com" },
]
description = "Fit dynamical models to single-cell data"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyro-ppl",
    "torch",
    "torchode",
    "numpy",
    "scipy",
    "matplotlib",
    "tqdm",
    "scikit-learn",
    "umap-learn",
    "anndata",
    "leidenalg",
    "openpyxl",
    "cmdstanpy",
    "arviz",
    "networkx",
    "seaborn",
    "scanpy",
]

[project.urls]
"Homepage" = "https://github.com/chvandorp/scdynsys"
"Bug Tracker" = "https://github.com/chvandorp/scdynsys"

[tool.setuptools]
include-package-data = true
packages = ["scdynsys", "scdynsys.vae"]